utils: Slightly refactor directory stashing and move it to its own file
This starts an effort to modularize the monolithic "Utils.kt" file.
